app/requirements.txt
Django>=4.2.6,<5.0.0
djangorestframework
djangorestframework-gis
flake8
pydocstyle
requests>=2.31.0,<3.0.0
psycopg2>=2.9.5,<3.0.0
django-json-widget>=1.1.1,<2.0.0
redis
django-redis
django-extensions
pytz